requirements.txt
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Werkzeug==2.2.2
gunicorn==21.2.0
numpy==1.22.4
pandas==1.4.4
tensorflow-cpu==2.12.0
scikit-learn==1.2.2
python-dateutil==2.8.2
